requirements.txt
streamlit
streamlit-drawable-canvas
pillow 
matplotlib
scikit-image